refactor(core): extract extraction envelope helpers; justified storage watchlist update
Structural ratchet response (#1520 tracker):
- extraction-persist.ts drops back under the 3000-line ceiling (2967) by
  extracting composeSalvagedExtractionEnvelope + withReservedMarkerTag
  into orchestration/extraction-envelope.ts - also deduplicates the
  salvage-note warn-log across all four write sites so a future site
  cannot forget it
- storage.ts watchlist 7010 -> 7062 is accepted with justification: the
  sealed-envelope write entry point (writeSealedMemory + SealedWriteExtras,
  issue #1989 PR2's core deliverable) belongs on StorageManager beside
  writeMemory; the +52 lines are that entry point and its contract docs
