+++ scripts/ratchet-baseline.json
@@ -7,7 +7,7 @@
       "packages/remnic-core/src/orchestrator.ts": 3980,
       "packages/remnic-core/src/cli.ts": 9429,
       "packages/remnic-core/src/access-service.ts": 5905,
-      "packages/remnic-core/src/storage.ts": 7010,
+      "packages/remnic-core/src/storage.ts": 7062,
       "packages/remnic-core/src/config.ts": 4561
     },
     "oversizedFileCount": 11,
